fix: small refactor to invitation details
